+++ test/fuse/read/testutils/FuseReadTest.h
@@ -16,12 +16,12 @@ public:
   void ReadFile(const char *filename, void *buf, size_t count, off_t offset);
   ReadError ReadFileReturnError(const char *filename, void *buf, size_t count, off_t offset);
 
-  ::testing::Action<int(int, void*, size_t, off_t)> ReturnSuccessfulRead =
+  ::testing::Action<size_t(int, void*, size_t, off_t)> ReturnSuccessfulRead =
     ::testing::Invoke([](int, void *, size_t count, off_t) {
       return count;
     });
 
-  ::testing::Action<int(int, void*, size_t, off_t)> ReturnSuccessfulReadRegardingSize(size_t filesize) {
+  ::testing::Action<size_t(int, void*, size_t, off_t)> ReturnSuccessfulReadRegardingSize(size_t filesize) {
     return ::testing::Invoke([filesize](int, void *, size_t count, off_t offset) {
       size_t ableToReadCount = std::min(count, (size_t)(filesize - offset));
       return ableToReadCount;
