+++ test/testutils/FuseTest.h
@@ -50,7 +50,7 @@ public:
   MOCK_METHOD2(fstat, void(int, struct ::stat*));
   MOCK_PATH_METHOD2(truncate, void, off_t);
   MOCK_METHOD2(ftruncate, void(int, off_t));
-  MOCK_METHOD4(read, int(int, void*, size_t, off_t));
+  MOCK_METHOD4(read, size_t(int, void*, size_t, off_t));
   MOCK_METHOD4(write, void(int, const void*, size_t, off_t));
   MOCK_METHOD1(flush, void(int));
   MOCK_METHOD1(fsync, void(int));
